pass windows test
Remove rmtree in windows test

+++ pyphs/tests/PHSSimulationPlotsTest.py
@@ -56,7 +56,8 @@ def plot_power_balance_rlc_with_split():
     simu.data.plot_powerbal(mode='single', show=False)
     simu.data.plot_powerbal(mode='multi', show=False)
 
-    shutil.rmtree(path)
+    if not os.name.lower().startswith('win'):
+        shutil.rmtree(path)
     return True
 
 
@@ -96,7 +97,8 @@ def plot_power_balance_nlcore_with_split():
     simu.data.plot_powerbal(mode='single', show=False)
     simu.data.plot_powerbal(mode='multi', show=False)
 
-    shutil.rmtree(path)
+    if not os.name.lower().startswith('win'):
+        shutil.rmtree(path)
     return True
 
 
@@ -147,7 +149,8 @@ def plot_rlc_with_split():
     simu.data.plot([('dtx', i) for i in range(dims.x())] +
                    [('dxH', i) for i in range(dims.x())], show=False)
 
-    shutil.rmtree(path)
+    if not os.name.lower().startswith('win'):
+        shutil.rmtree(path)
     return True
